feat(searchbox): Add `useSearchBox` hook
This hook is a companion to the SearchBox component. It allows for consumers to build their
searchbox component from scratch instead of using the out of the box SearchBox component.
