+++ src/search/search-box/search-box.tsx
@@ -1,20 +1,9 @@
 import cn from 'classnames';
-import {
-  ChangeEvent,
-  ForwardedRef,
-  forwardRef,
-  ReactElement,
-  useContext,
-  useEffect,
-  useState,
-  useCallback,
-} from 'react';
+import { ForwardedRef, ReactElement, forwardRef } from 'react';
 
-import { debounce } from '../../utils/debounce';
-import { SearchContext } from '../context/search.context';
-import { useProductSearch } from '../hooks/productSearch.hook';
 import './search-box.scss';
 
+import { useSearchBox } from './search-box.hook';
 import type { SearchBoxProps } from './search-box.types';
 
 /**
@@ -29,18 +18,12 @@ import type { SearchBoxProps } from './search-box.types';
  *
  * // Set the account and catalog configuration
  * const config: Configuration = {
- *   account_id: 1234,
- *   domain_key: 'example_com',
+ *    ...
  * };
  *
  * // Set up the search parameters
  * const searchOptions: ProductSearchOptions = {
- *   q: 'Generic Metal Pants',
- *   fl: 'pid,title,description,brand,price,thumb_image',
- *   start: 0,
- *   rows: 10,
- *   url: 'http://example.com',
- *   _br_uid_2: 'someCookieId',
+ *   ...
  * };
  *
  * export default function MyCustomComponent() {
@@ -51,33 +34,35 @@ import type { SearchBoxProps } from './search-box.types';
  *        searchOptions={searchOptions}
  *        debounceDelay={300}
  *        className="test"/>
+ *
+ *      <SearchContext.Consumer>
+ *        {({ searchResponse }) =>
+ *          searchResponse?.response?.docs?.map((result) => {
+ *            return (
+ *              <div>
+ *                <h2>{result.title}</h2>
+ *                <p>{result.description}</p>
+ *              </div>
+ *            );
+ *          })
+ *        }
+ *      </SearchContext.Consumer>
  *    </SearchContextProvider>
  *  )
- *   return <SearchBox />;
  * }
+ *
+ *
  * ```
  */
 export const SearchBox = forwardRef(
   (props: SearchBoxProps, forwardedRef: ForwardedRef<HTMLInputElement> | null): ReactElement => {
     const { children, className, configuration, searchOptions, debounceDelay, ...rest } = props;
-    const searchContext = useContext(SearchContext);
-    const [query, setQuery] = useState<string>('');
-    const { response } = useProductSearch(query, configuration, searchOptions);
-
-    useEffect(() => {
-      searchContext.setSearchResponse(response);
-    }, [searchContext, response]);
-
-    const debouncedSetQuery = debounce((event: ChangeEvent<HTMLInputElement>) => {
-      setQuery(event.target.value);
-    }, debounceDelay ?? 500);
-
-    const onChange = useCallback(debouncedSetQuery, [debouncedSetQuery]);
+    const { changeHandler } = useSearchBox(props);
 
     return (
       <input
         {...rest}
-        onChange={onChange}
+        onChange={changeHandler}
         className={cn('lcui-search-box', className)}
         ref={forwardedRef}
       />
